requirements.txt
ultralytics>=8.0.0
torch
torchvision
torchaudio
opencv-python-headless
streamlit
pyyaml
tqdm
numpy
